完成后台后端language模块功能实现

+++ oj-admin/src/main/java/com/creative/ojadmin/controller/language/param/CreateLanguageParam.java
@@ -12,7 +12,7 @@ public class CreateLanguageParam {
 
     private String label;
 
-    private long value;
+    private int value;
 
     private boolean enable;
